requirements.txt
fastapi
uvicorn
pydantic
PyJWT
passlib[bcrypt]
pytest

httpx
python-multipart
ollama
requests
chromadb
sentence-transformers
